Add test website
This is the project testbed and may serve as a foundation for some
usable product.  It is similar to the demo on the Raphael SketchPad
website.  NB the use of jQuery Migrate, as SketchPad uses
jQuery.browser(), which was removed in jQuery 1.9.  Also, be sure to
use $(document).ready() (or something similar) to ensure the page has
loaded before calling Raphael.sketchpad().