Add preliminary implementation of "point scattering" algorithm
Added the algorithm that will evenly distribute some number
of points across a path.  This is not intended to be a path
simplification algorithm, but rather just to sample points, for
use in the shape context algorithm.  This has not yet been
tested, but I wanted to commit what I have.