Finish shapeContextMatch and include Hungarian algorithm
It seems to work to some extent!  Certainly not perfected,
and the return value is not in the range 0 to 100, but rather
an error value is returned (e.g. 0 is a perfect match).  More
testing and tweaks will be required, but everything seems to
right in terms of the shape context algorithm I was reading about.
Also, the Hungarian algorithm piece was a big enough task to
tackle that I made it its own project, repository and all.  As not to
require any additional files just yet, I pasted it in here.  That was
a beast of its own.