Fix distributePointsAcrossPath bug
This is the algorithm I originally intended.  It is used by
scatterPoints, which is used by both Betim's and my algorithms.
The bug was a - sign instead of a + sign when calculating the
location of a new point off to a certain angle.  Now the shapeContext
algorithm works a whole lot better with this correct algorithm!